refactor: 에러메세지 분류
