+++ docs/README.md
@@ -6,6 +6,7 @@
 - [ ] 당첨번호를 입력받는 기능을 구현한다.
 - [ ] 보너스 번호를 입력받는 기능을 구현한다.
 - [ ] 당첨번호와 구입한 번호를 비교하여 등수를 판별하는 기능을 구현한다.
+- [ ] 일치한 번호의 개수가 5개일때 보너스 번호도 검사하는 기능을 구현한다.
 - [ ] 판별한 등수에 따라 얻는 수익금을 계산하는 기능을 구현한다.
 - [ ] 수익금과 구입금액을 비교하여 수익률을 계산하는 기능을 구현한다.
